feat: extract background loadWellKnownChains and updateDatabases
